requirements.txt
beautifulsoup4==4.11.1
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
plotly==5.10.0
requests==2.28.1
streamlit==1.11.1
